refactor : move exception related classes into exception package
