+++ src/main/java/coz/weavon/common/application/util/AuthUtils.java
@@ -1,6 +1,6 @@
 package coz.weavon.common.application.util;
 
-import coz.weavon.common.application.model.exception.AuthException;
+import coz.weavon.exception.model.AuthException;
 import coz.weavon.core.auth.domain.model.AuthUser;
 import java.util.Objects;
 import lombok.experimental.UtilityClass;
