+++ src/main/java/coz/weavon/core/auth/infrastructure/adapter/AuthUserRestAdapter.java
@@ -1,6 +1,6 @@
 package coz.weavon.core.auth.infrastructure.adapter;
 
-import coz.weavon.common.application.model.exception.BusinessException;
+import coz.weavon.exception.model.BusinessException;
 import coz.weavon.common.infrastructure.model.Adapter;
 import coz.weavon.core.auth.application.adapter.AuthUserAdapter;
 import coz.weavon.core.auth.domain.model.AuthUser;
