+++ src/main/java/coz/weavon/core/user/application/model/condition/UserSearchCondition.java
@@ -2,7 +2,7 @@ package coz.weavon.core.user.application.model.condition;
 
 import com.querydsl.core.types.dsl.BooleanExpression;
 import coz.weavon.common.application.model.condition.RestCondition;
-import coz.weavon.common.application.model.exception.BusinessException;
+import coz.weavon.exception.model.BusinessException;
 import coz.weavon.core.user.application.model.command.UserSearchCommand;
 import coz.weavon.core.user.infrastructure.model.QUserEntity;
 import java.util.List;
