+++ src/main/java/coz/weavon/exception/model/AuthException.java
@@ -1,4 +1,4 @@
-package coz.weavon.common.application.model.exception;
+package coz.weavon.exception.model;
 
 public class AuthException extends RestException {
